% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcPercChangeMCMC.R
\name{calcPercChangeMCMC}
\alias{calcPercChangeMCMC}
\title{calcPercChangeMCMC}
\usage{
calcPercChangeMCMC(vec.in, model.in = NULL, perc.change.bm = -25,
  na.skip = FALSE, out.type = "short", mcmc.plots = FALSE,
  convergence.check = TRUE)
}
\arguments{
\item{vec.in}{vector with numeric values}

\item{model.in}{if NULL, use the BUGS code in the built in function trend.bugs.1()}

\item{perc.change.bm}{benchmark for Prob(Decl>BM), default = -25}

\item{na.skip}{if TRUE, skip the calculations when vec.in contains any NA}

\item{out.type}{"short" or "long". 
"short" gives summary table of posterior plus "PercChange" and "ProbDecl"
"long" also includes the full mcmc samples and the jags output object}

\item{mcmc.plots}{if true, create the standard series of MCMC diagnostic plots 
Note that these are printed to the default device 
(i.e. need to external wrap the function call inside a pdf /dev.off call)
To get a plot of the model fit, run this function with out.type = "long",
and then use plot.trend.fit().}

\item{convergence.check}{if TRUE, do an automated convergence check}
}
\description{
MCMC version of trend metric calculation (replaces old version based on PBSModelling and Metropolis algorithm)
}
\examples{
pdf("ProbDecl_Fits.pdf",onefile=TRUE,height=8.5, width=11)
vec.use = c(12,10,14,7,13,5,8,3,4,7,6,5)
slope.mcmc.fit <- calcPercChangeMCMC(vec.in= vec.use,model.in = trend.bugs.1 , 
                                    perc.change.bm = -25 , na.skip=FALSE, 
                                    out.type = "long", mcmc.plots = TRUE)
dev.off()
}
\keyword{MCMC,}
\keyword{slope,}
\keyword{trend}
